Add basic unit tests to service layer
Add two classes, to test UserService and Category service.
TransactionService contains almost no logic, so it doesn't need
to be tested. I feel like these tests are useless, because the
logic is very simple, and it won't be affected by changes in
another parts of system. But it's fine for practice.